Update Jan 20, 2016 It seems that Polan had overseen a few things when constructing this API. It always loads everything it can load. Not really a problem for most data as that is just text, but it also loads all trophy images regardless if that action was actually needed. Now THIS slows down login time big deal. Now the API does not do that any more by default, however if you still want the API to do this, just start you BlitzMax program after the 'strict', 'framework' and 'import' sessions with the line 'GJImageSkip=False'.
If it still slows down after that, we might conclude there is a server issue.

Now, here I was thinking I found a GameJolt bug while the true evil was actually of the origal creator's doing. Well, I'm glad that is out of the way now, but I think I should apologize to the guys at GJ for this. Sorry!

Oh yeah, not only will this update make your game faster. It also prevents you from wasting a lot of memory too (at least it did in *my* case).

Having said that, I can
- close #1
- close Tricky1975/Star-Story#309
- close gamejolt/issue-tracker#438 (if that is possible since that repo isn't mine, but it was my issue, so we'll see what happens)

Anyway, everything's fine so it seems.